moved common figure arguments to `docs.py`. Added examples to figures.
